+++ ThumbnailWindow.cpp
@@ -60,9 +60,8 @@ ThumbnailWindow::ThumbnailWindow(HWND targetWindow, RECT cropRect, bool showTitl
     }
 
     m_hostWindow = CreateWindowExW(
-        exStyle, ClassName, L"ZenCrop - Thumbnail", style,
-        cropRect.left - BorderWidth, cropRect.top - BorderWidth,
-        windowWidth, windowHeight,
+        exStyle, ClassName, L"Thumbnail", style,
+        cropRect.left, cropRect.top, windowWidth, windowHeight,
         nullptr, nullptr, GetModuleHandleW(nullptr), this
     );